Rewrote Makefile to perform out-of-source build.
All build artifacts are now put into the `build` folder. The `src`
folder will remain untouched.

We define three different build targets: "release", "debug", and
"memtest". The latter is used to detect memory leaks at run time. It
detects a number of them with the current code.

Now that the source code for tools was moved out, we can assume that
all source files must be compiled and linked into a single executable.
This simplifies the structure of the `Makefile`.