[wip] Replace FFTW with HeFFTe as the FFT library.
This is work in progress. The code, as is, segfaults. The memory layout
of the FFT components as produced by HeFFTe is certainly different from
the way FFTW does it, and much of the Pinocchio code relies on that
assumption, for example when it computes derivatives.

Namely, when using FFTW, we asked it to transpose the dimensions in
Fourier space, for performance reasons. We need to figure out how to
get HeFFTe to do the same. It's also not clear why HeFFTe doesn't
determine the memory size of the output box by itself. As it is now,
it just relies on our input, which is most likely wrong.