Add support for import and export of initial density
Until now, initial conditions could be set via a "white-noise file",
using the `-DWHITENOISE` compile-time option implemented in
`ReadWhiteNoise.c`. But that procedure imposes a power spectrum.

This change adds support for importing and exporting the initial mass
overdensity (initial density for short) directly. We can thus compare
Pinocchio results more easily with N-body simulations that have a
similar import/export feature.

See the doc-comment in `initial_density.c` as to the format of the file.
To import the density, add the option `InitialDensityFile` to the input
parameter file and specify the file name as its value. To export the
density generated from the random seed, pass `--write-initial-density`
as a command-line argument to the Pinocchio executable. Optionally, it
can be followed by a file name, which defaults to `initial_density.dat`.